Add the ability use interact and solid box checks
